// web/src/gen/proto/medication/v1/medication_entry_pb.ts
// @generated by protoc-gen-es v2.12.1 with parameter "target=ts"
// @generated from file proto/medication/v1/medication_entry.proto (package proto.medication.v1, syntax proto3)
/* eslint-disable */

import type { GenEnum, GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { enumDesc, fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file proto/medication/v1/medication_entry.proto.
 */
export const file_proto_medication_v1_medication_entry: GenFile = /*@__PURE__*/
  fileDesc("Cipwcm90by9tZWRpY2F0aW9uL3YxL21lZGljYXRpb25fZW50cnkucHJvdG8SE3Byb3RvLm1lZGljYXRpb24udjEiiwEKCk1lZGljYXRpb24SCgoCaWQYASABKA0SDAoEbmFtZRgCIAEoCRIaChJyZWNvbW1lbmRlZF9kb3NhZ2UYAyABKAISRwoSZG9zYWdlX21lYXN1cmVtZW50GAQgASgOMisucHJvdG8ubWVkaWNhdGlvbi52MS5NRURJQ0FUSU9OTUVBVVNFUkVNRU5UKn0KFk1FRElDQVRJT05NRUFVU0VSRU1FTlQSJgoiTUVESUNBVElPTk1FQVVTRVJFTUVOVF9VTlNQRUNJRklFRBAAEh0KGU1FRElDQVRJT05NRUFVU0VSRU1FTlRfTUcQARIcChhNRURJQ0FUSU9OTUVBVVNFUkVNRU5UX0cQAmIGcHJvdG8z");

/**
 * @generated from message proto.medication.v1.Medication
 */
export type Medication = Message<"proto.medication.v1.Medication"> & {
  /**
   * @generated from field: uint32 id = 1;
   */
  id: number;

  /**
   * @generated from field: string name = 2;
   */
  name: string;

  /**
   * @generated from field: float recommended_dosage = 3;
   */
  recommendedDosage: number;

  /**
   * @generated from field: proto.medication.v1.MEDICATIONMEAUSEREMENT dosage_measurement = 4;
   */
  dosageMeasurement: MEDICATIONMEAUSEREMENT;
};

/**
 * Describes the message proto.medication.v1.Medication.
 * Use `create(MedicationSchema)` to create a new message.
 */
export const MedicationSchema: GenMessage<Medication> = /*@__PURE__*/
  messageDesc(file_proto_medication_v1_medication_entry, 0);

/**
 * @generated from enum proto.medication.v1.MEDICATIONMEAUSEREMENT
 */
export enum MEDICATIONMEAUSEREMENT {
  /**
   * @generated from enum value: MEDICATIONMEAUSEREMENT_UNSPECIFIED = 0;
   */
  MEDICATIONMEAUSEREMENT_UNSPECIFIED = 0,

  /**
   * @generated from enum value: MEDICATIONMEAUSEREMENT_MG = 1;
   */
  MEDICATIONMEAUSEREMENT_MG = 1,

  /**
   * @generated from enum value: MEDICATIONMEAUSEREMENT_G = 2;
   */
  MEDICATIONMEAUSEREMENT_G = 2,
}

/**
 * Describes the enum proto.medication.v1.MEDICATIONMEAUSEREMENT.
 */
export const MEDICATIONMEAUSEREMENTSchema: GenEnum<MEDICATIONMEAUSEREMENT> = /*@__PURE__*/
  enumDesc(file_proto_medication_v1_medication_entry, 0);